requirements.txt
streamlit>=1.28.0
google-genai>=0.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
opencv-python>=4.8.0
